rename non-suid -> rootless
