+++ Makefile
@@ -22,9 +22,9 @@ all: squashfs-mount
 
 %.o: %.c
 	$(CC) $(CFLAGS) $(SQFS_CFLAGS) $(SQUASHFS_MOUNT_CFLAGS) -c -o $@ $<
-squashfs-mount.o: VERSION non-suid.h
+squashfs-mount.o: VERSION rootless.h
 
-squashfs-mount: squashfs-mount.o non-suid.o
+squashfs-mount: squashfs-mount.o rootless.o
 		$(CC) $^ $(LDFLAGS) $(SQUASHFS_MOUNT_LDFLAGS) -o $@
 
 install: squashfs-mount
@@ -35,9 +35,9 @@ install-suid: install
 	chown root:root $(DESTDIR)$(bindir)/squashfs-mount
 	chmod u+s $(DESTDIR)$(bindir)/squashfs-mount
 
-rpm: squashfs-mount.c non-suid.c non-suid.h VERSION LICENSE Makefile
+rpm: squashfs-mount.c rootless.c rootless.h VERSION LICENSE Makefile
 	./generate-rpm.sh -b $@
 	$(RPMBUILD) -bs --define "_topdir $@" $@/SPECS/squashfs-mount.spec
 
 clean:
-	rm -rf squashfs-mount squashfs-mount.o non-suid.o rpm
+	rm -rf squashfs-mount squashfs-mount.o rootless.o rpm
